spat_llava/requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
anaconda-clean @ file:///opt/conda/conda-bld/anaconda-clean_1652723720992/work
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
blessed==1.20.0
ccimport==0.4.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cumm-cu120==0.4.11
cycler==0.12.1
deepspeed==0.12.6
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.1
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.4
fire==0.6.0
flash-attn==2.5.7
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.25.0
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lark==1.2.2
# Editable Git install with no remote (llava==1.2.2.post1)
-e /home/aiops/wangzh/llava-spat
loralib==0.1.2
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
ninja==1.11.1.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.1
packaging==24.0
pandas==2.2.3
pccm==0.4.16
peft==0.10.0
pillow==10.3.0
portalocker==2.10.1
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pybind11==2.13.6
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.1
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.45.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spconv-cu120==2.3.6
starlette==0.37.2
svgwrite==1.4.3
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.4.0
tokenizers==0.15.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.5
transformers==4.37.2
triton==2.1.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wandb==0.16.6
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
xxhash==3.4.1
yarl==1.9.4
